IndexedDB/tests/submissions/Opera/idbcursor-source.htm

<!DOCTYPE html>
<title>IDBCursor.source</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support.js"></script>

<script>
    setup({ explicit_done: true });

    var db;
    var open_rq = indexedDB.open('testdb-' + new Date().getTime());
    open_rq.onupgradeneeded = function(e) {
        db = e.target.result;
        var objStore = db.createObjectStore("test");
        objStore.createIndex("index", "");

        objStore.add("data",  1);
        objStore.add("data2", 2);
    };

    function cursor_source(stringified_object, cursor_rq) {
        var cursor;

        cursor_rq.onsuccess = this.step_func(function(e) {
            if (!e.target.result) {
                return;
            }
            cursor = e.target.result;
            assert_readonly(cursor, 'source');

            // Direct try
            assert_true(cursor.source instanceof Object, "source isobject");
            assert_equals(cursor.source + "", stringified_object, "source");
            assert_idl_attribute(cursor.source, 'name');

            // Indirect try
            setTimeout(this.step_func(function() {
                assert_true(cursor.source instanceof Object, "source isobject");
                assert_equals(cursor.source + "", stringified_object, "source");
                assert_idl_attribute(cursor.source, 'name');
            }), 0);

            cursor.continue();
        });

        cursor_rq.transaction.oncomplete = this.step_func(function(e) {
            // A try when transaction is inactive
            assert_true(cursor.source instanceof Object, "source isobject");
            assert_equals(cursor.source + "", stringified_object, "source");
            assert_idl_attribute(cursor.source, 'name');

            // A last little try
            setTimeout(this.step_func(function() {
                assert_true(cursor.source instanceof Object, "source isobject");
                assert_equals(cursor.source + "", stringified_object, "source");
                assert_idl_attribute(cursor.source, 'name');

                this.done();
            }), 5);
         });
    }

    open_rq.onsuccess = function() {
        async_test(document.title + ' - IDBObjectStore').step(function() {
            cursor_source.call(this, "[object IDBObjectStore]", db.transaction("test")
                                                       .objectStore("test")
                                                       .openCursor());
        });

        async_test(document.title + ' - IDBIndex').step(function() {
            cursor_source.call(this, "[object IDBIndex]", db.transaction("test")
                                                 .objectStore("test")
                                                 .index("index")
                                                 .openCursor());
        });

        done();
    };

</script>

<div id="log"></div>
